+++ Accounts/PresentationLayer/Views/Forms/CashDisbursementJournalEntry.vb
@@ -169,7 +169,7 @@ Namespace PresentationLayer.Views.Forms
 
         Public Property PayeeIdNo As Int32? Implements ICashDisbursementJournalView.PayeeIdNo
             Get
-                Return cboPayeeIdNo.GetNullableValue(Of Int16)
+                Return cboPayeeIdNo.GetNullableValue(Of Int32)
             End Get
             Set
                 cboPayeeIdNo.SetValue(Value)
