+++ Accounts/PresentationLayer/Views/Forms/CheckDisbursementJournalEntry.vb
@@ -194,7 +194,7 @@ Namespace PresentationLayer.Views.Forms
 
         Public Property PayeeIdNo As Int32? Implements ICheckDisbursementJournalView.PayeeIdNo
             Get
-                Return cboPayeeIdNo.GetNullableValue(Of Int16)
+                Return cboPayeeIdNo.GetNullableValue(Of Int32)
             End Get
             Set
                 cboPayeeIdNo.SetValue(Value)
